+++ tests/testthat/test-planning.R
@@ -4,215 +4,163 @@ context("Planning")
 
 test_that(desc = "Frequentist Poisson 1% materiality", {
   jfaRes <- planning(materiality = 0.01, confidence = 0.95, expectedError = 0, likelihood = "poisson")
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 300)
   expect_equal(jfaRes$expectedSampleError, 0)
 })
 
 test_that(desc = "Frequentist Poisson 5% materiality", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0, likelihood = "poisson")
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 60)
   expect_equal(jfaRes$expectedSampleError, 0)
 })
 
 test_that(desc = "Frequentist Poisson 5% materiality 2.5% errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "poisson")
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 231)
-  expect_equal(jfaRes$expectedSampleError, 5.78)
+  expect_equal(jfaRes$expectedSampleError, 5.78, tolerance = 0.001)
 })
 
-test_that(desc = "Frequentist Poisson 5% materiality 2 errors", {
+test_that(desc = "Frequentist Poisson 5% materiality 2% errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 2, likelihood = "poisson")
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 126)
   expect_equal(jfaRes$expectedSampleError, 2)
 })
 
 test_that(desc = "Bayesian Poisson 5% materiality 2.5% errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "poisson", prior = TRUE)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 231)
-  expect_equal(jfaRes$expectedSampleError, 5.78)
+  expect_equal(jfaRes$expectedSampleError, 5.78, tolerance = 0.001)
 })
 
 test_that(desc = "Bayesian Poisson 5% materiality 2.5% errors custom prior", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "poisson", prior = TRUE, kPrior = 1, nPrior = 7)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 285)
-  expect_equal(jfaRes$expectedSampleError, 7.12)
+  expect_equal(jfaRes$expectedSampleError, 7.12, tolerance = 0.001)
 })
 
 test_that(desc = "Frequentist binomial 1% materiality", {
   jfaRes <- planning(materiality = 0.01, confidence = 0.95, expectedError = 0, likelihood = "binomial")
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 299)
   expect_equal(jfaRes$expectedSampleError, 0)
 })
 
 test_that(desc = "Frequentist binomial 5% materiality", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0, likelihood = "binomial")
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 59)
   expect_equal(jfaRes$expectedSampleError, 0)
 })
 
 test_that(desc = "Frequentist binomial 5% materiality 2.5% errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "binomial")
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 234)
-  expect_equal(jfaRes$expectedSampleError, 6)
+  expect_equal(jfaRes$expectedSampleError, 6, tolerance = 0.001)
 })
 
 test_that(desc = "Frequentist binomial 5% materiality 2 errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 2, likelihood = "binomial")
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 124)
   expect_equal(jfaRes$expectedSampleError, 2)
 })
 
 test_that(desc = "Bayesian binomial 5% materiality 2.5% errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "binomial", prior = TRUE)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 220)
-  expect_equal(jfaRes$expectedSampleError, 5.5)
+  expect_equal(jfaRes$expectedSampleError, 5.5, tolerance = 0.001)
 })
 
 test_that(desc = "Bayesian binomial 5% materiality 2.5% errors custom prior", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "binomial", prior = TRUE, kPrior = 1, nPrior = 7)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 273)
 })
 
 test_that(desc = "Frequentist hypergeometric 1% materiality", {
   jfaRes <- planning(materiality = 0.01, confidence = 0.95, expectedError = 0, likelihood = "hypergeometric", N = 1000)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 258)
   expect_equal(jfaRes$expectedSampleError, 0)
 })
 
 test_that(desc = "Frequentist hypergeometric 5% materiality", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0, likelihood = "hypergeometric", N = 1000)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 57)
   expect_equal(jfaRes$expectedSampleError, 0)
 })
 
 test_that(desc = "Frequentist hypergeometric 5% materiality 2.5% errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "hypergeometric", N = 1000)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 197)
-  expect_equal(jfaRes$expectedSampleError, 5)
+  expect_equal(jfaRes$expectedSampleError, 5, tolerance = 0.001)
 })
 
 test_that(desc = "Frequentist hypergeometric 5% materiality 2 errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 2, likelihood = "hypergeometric", N = 1000)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 119)
   expect_equal(jfaRes$expectedSampleError, 2)
 })
 
 test_that(desc = "Bayesian hypergeometric 5% materiality 2.5% errors", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "hypergeometric", N = 1000, prior = TRUE)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 160)
-  expect_equal(jfaRes$expectedSampleError, 4)
+  expect_equal(jfaRes$expectedSampleError, 4, tolerance = 0.001)
 })
 
 test_that(desc = "Bayesian hypergeometric 5% materiality 2.5% errors custom prior", {
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.025, likelihood = "hypergeometric", N = 1000, prior = TRUE, kPrior = 1, nPrior = 7)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 193)
-  expect_equal(jfaRes$expectedSampleError, 5)
+  expect_equal(jfaRes$expectedSampleError, 5, tolerance = 0.001)
 })
 
 test_that(desc = "Bayesian binomial 2% precision 5% materiality 2.5% errors median prior", {
   jfaPrior <- auditPrior(materiality = 0.05, confidence = 0.95, method = "median", expectedError = 0.025)
   jfaRes <- planning(materiality = 0.05, minPrecision = 0.02, confidence = 0.95, expectedError = 0.025, likelihood = "binomial", N = 1000, prior = jfaPrior)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 285)
-  expect_equal(jfaRes$expectedSampleError, 7.12)
+  expect_equal(jfaRes$expectedSampleError, 7.12, tolerance = 0.001)
 })
 
 test_that(desc = "Bayesian binomial 5% materiality 1% errors hypotheses prior", {
   jfaPrior <- auditPrior(materiality = 0.05, confidence = 0.95, method = "hypotheses", pHmin = 0.7)
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.01, likelihood = "binomial", N = 1000, prior = jfaPrior)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 55)
-  expect_equal(jfaRes$expectedSampleError, 0.55)
+  expect_equal(jfaRes$expectedSampleError, 0.55, tolerance = 0.001)
 })
 
 test_that(desc = "Bayesian poisson 5% materiality 1% errors factor prior", {
   jfaPrior <- auditPrior(materiality = 0.05, confidence = 0.95, method = "factor", factor = 0.6, sampleN = 58, sampleK = 0, expectedError = 0.025)
   jfaRes <- planning(materiality = 0.05, confidence = 0.95, expectedError = 0.01, likelihood = "binomial", N = 1000, prior = jfaPrior)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 36)
-  expect_equal(jfaRes$expectedSampleError, 0.36)
+  expect_equal(jfaRes$expectedSampleError, 0.36, tolerance = 0.001)
 })
 
 test_that(desc = "Bayesian poisson 3% materiality 1.3% errors sample prior", {
   jfaPrior <- auditPrior(materiality = 0.03, confidence = 0.95, method = "sample", sampleN = 58, sampleK = 0, expectedError = 0.025)
   jfaRes <- planning(materiality = 0.03, confidence = 0.95, expectedError = 0.013, likelihood = "binomial", N = 1000, prior = jfaPrior)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 140)
-  expect_equal(jfaRes$expectedSampleError, 1.82)
+  expect_equal(jfaRes$expectedSampleError, 1.82, tolerance = 0.001)
 })
 
 test_that(desc = "Frequentist poisson 3% materiality 1.3% errors 5% min precision", {
   jfaRes <- planning(materiality = 0.03, confidence = 0.95, expectedError = 0.013, minPrecision = 0.05, likelihood = "binomial", N = 1000)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 303)
-  expect_equal(jfaRes$expectedSampleError, 4)
+  expect_equal(jfaRes$expectedSampleError, 4, tolerance = 0.001)
 })
 
 # jfa version 0.2.0
 
 test_that(desc = "Bayesian poisson 3% materiality 1.3% errors 5% min precision standard prior", {
   jfaRes <- planning(materiality = 0.03, confidence = 0.95, expectedError = 0.013, minPrecision = 0.05, likelihood = "binomial", N = 1000, prior = T)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 293)
-  expect_equal(jfaRes$expectedSampleError, 3.81)
+  expect_equal(jfaRes$expectedSampleError, 3.81, tolerance = 0.001)
 })
 
 test_that(desc = "Bayesian poisson 100% materiality 0% errors 2% min precision standard prior", {
-  jfaRes <- planning(confidence = 0.95, minPrecision = 0.02, likelihood = "poisson", prior = T)
-  print(jfaRes)
-  plot(jfaRes)
+  jfaRes <- planning(confidence = 0.95, minPrecision = 0.02, likelihood = "poisson", prior = TRUE)
   expect_equal(jfaRes$sampleSize, 150)
   expect_equal(jfaRes$expectedSampleError, 0)
 })
 
 test_that(desc = "Bayesian poisson 100% materiality 1% errors 2% min precision standard prior", {
   jfaRes <- planning(confidence = 0.95, minPrecision = 0.02, expectedError = 0.01, likelihood = "poisson", prior = T)
-  print(jfaRes)
-  plot(jfaRes)
   expect_equal(jfaRes$sampleSize, 220)
-  expect_equal(jfaRes$expectedSampleError, 2.2)
+  expect_equal(jfaRes$expectedSampleError, 2.2, tolerance = 0.001)
 })
 
 # jfa version 0.3.0
@@ -222,4 +170,25 @@ test_that(desc = "Bayesian poisson 100% materiality 1% errors 2% min precision s
 # No changes to be tested
 
 # jfa version 0.4.0
-# No changes to be tested
+
+test_that(desc = "Expected Bayes factors for zero expected errors", {
+  jfaRes <- planning(confidence = 0.95, materiality = 0.02, expectedError = 0, likelihood = "poisson", prior = T)
+  expect_equal(jfaRes$hypotheses$expectedBf, 20.08554, tolerance = 0.001)
+  
+  jfaRes <- planning(confidence = 0.95, materiality = 0.02, expectedError = 0.01, likelihood = "binomial", prior = T)
+  expect_equal(jfaRes$hypotheses$expectedBf, 19.64692, tolerance = 0.001)
+  
+  jfaRes <- planning(confidence = 0.95, materiality = 0.02, expectedError = 0.01, likelihood = "hypergeometric", prior = T, N = 1000)
+  expect_equal(jfaRes$hypotheses$expectedBf, 5.001883, tolerance = 0.001)
+})
+
+test_that(desc = "Expected Bayes factors for expected errors > 0", {
+  jfaRes <- planning(confidence = 0.95, materiality = 0.02, expectedError = 0.01, likelihood = "poisson", prior = T)
+  expect_equal(jfaRes$hypotheses$expectedBf, 19.8358, tolerance = 0.001)
+  
+  jfaRes <- planning(confidence = 0.95, materiality = 0.02, expectedError = 0.01, likelihood = "binomial", prior = T)
+  expect_equal(jfaRes$hypotheses$expectedBf, 19.64692, tolerance = 0.001)
+  
+  jfaRes <- planning(confidence = 0.95, materiality = 0.02, expectedError = 0.01, likelihood = "hypergeometric", prior = T, N = 1000)
+  expect_equal(jfaRes$hypotheses$expectedBf, 5.001883, tolerance = 0.001)
+})